move function into main_package
